added user login route for clients.

+++ src/utils/jwt.rs
@@ -9,15 +9,9 @@ use crate::errors::AppError;
 
 use super::{get_env, tools::keygen::BEARER_KEY};
 
-// pub static JWT_EXPIRY: i64 = days_to_seconds(1);
-
-// const fn days_to_seconds(days: i64) -> i64 {
-//     3600 * 24 * days
-// }
-
 #[derive(Deserialize, Serialize)]
 pub struct Claims {
-    pub sub: u64,
+    pub sub: i32,
     pub exp: i64,
 }
 
@@ -51,7 +45,7 @@ fn extract_jwt(header_value: &HeaderValue) -> Result<String, AppError> {
     ))
 }
 
-pub(crate) fn create_jwt(user_id: &u64, secret: &[u8], exp: i64) -> Result<String, AppError> {
+pub(crate) fn create_jwt(user_id: &i32, secret: &[u8], exp: i64) -> Result<String, AppError> {
     let exp = Utc::now()
         .checked_add_signed(chrono::Duration::seconds(exp))
         .expect("Invalid timestamp")
